Removed old class GlobalUrls, updated the tests
